Inline streaming invoke mode implementation
